fix: use bfloat16 for MXFP8 scale decompression
Convert E8M0 scale to bfloat16 instead of float32 for consistency
with model weight dtype. Wrap decompressed scale as nn.Parameter
for proper model loading.

+++ src/compressed_tensors/compressors/mxfp8/base.py
@@ -67,10 +67,10 @@ class MXFP8QuantizationCompressor(NaiveQuantizationCompressor):
         """
         state_dict = state_dict.copy()
 
-        # Convert E8M0 scale back to float
+        # Convert E8M0 scale back to bfloat16 for consistency with model dtype
         scale = state_dict["weight_scale"]
         scale_exp = scale.to(torch.int32) - 127
-        scale_float = 2.0 ** scale_exp.to(torch.float32)
+        scale_float = 2.0 ** scale_exp.to(torch.bfloat16)
         state_dict["weight_scale"] = scale_float
 
         return NaiveQuantizationCompressor.decompress(state_dict, scheme)
